Static columns and new empty mobile view
