+++ src/pages/registration/FilesAndLicensePanel.tsx
@@ -21,7 +21,7 @@ import {
   getChannelRegisterJournalUrl,
   getChannelRegisterPublisherUrl,
 } from '../public_registration/PublicPublicationContext';
-import { FileList } from './FileList';
+import { FileList } from './files/FileList';
 import { FileUploader } from './files_and_license_tab/FileUploader';
 import { DoiField } from './resource_type_tab/components/DoiField';
