+++ src/pages/registration/files/MobileFormattedFiles.tsx
@@ -1,0 +1,5 @@
+import { Box } from '@mui/material';
+
+export const MobileFormattedFiles = () => {
+  return <Box>test</Box>;
+};
